Complete task of getting Elsa to compile and run on modern GCC, specifically GCC-5.2.1 on x86_64.
Some of the changes involve removing the last uses of the pre-C++98
header files that ended with ".h".  Others suppress warnings from
GCC.

There are also two fixes for operator precedence bugs, one in
cc_type.cc and one in stdconv.cc, both found because of the GCC
warnings.  I haven't created any new tests to go with these fixes.

+++ ccparse.cc
@@ -5,7 +5,7 @@
 #include "astvisit.h"     // ASTVisitorEx
 #include "trace.h"        // TRACE
 
-#include <iostream.h>     // cout
+#include "sm-iostream.h"  // cout
 
 
 // ----------------------- ParseEnv -----------------------
@@ -38,7 +38,7 @@ SimpleTypeId ParseEnv::uberSimpleType(SourceLoc loc, UberModifiers m)
   m = (UberModifiers)(m & UM_TYPEKEYS);
 
   // implement cppstd Table 7, p.109
-  switch (m) {
+  switch ((int /*suppress GCC warning*/)m) {
     case UM_CHAR:                         return ST_CHAR;
     case UM_UNSIGNED | UM_CHAR:           return ST_UNSIGNED_CHAR;
     case UM_SIGNED | UM_CHAR:             return ST_SIGNED_CHAR;
